play audio using WASAPI
same model as for core audio, we continuously fill a buffer
of audio data.

the buffer clock is built from the difference between the time reported by
the API as coming from the speakers, and the number of rendered frames.

no attempt is being made at cleaning up the loaded resources at exit.